+++ services/import/matchupCategoryService.js
@@ -1,7 +1,7 @@
 import moment from 'moment';
-import pool from './db.js';
+import pool from '../db.js';
 import yahooApiService from './yahooApiService.js';
-import viewService from './viewService.js';
+import viewService from '../viewService.js';
 import seasonService from './seasonService.js';
 import gameCodeService from './gameCodeService.js';
 import seasonWeekService from './seasonWeekService.js';
