+++ services/import/rosterPositionService.js
@@ -1,4 +1,4 @@
-import pool from './db.js';
+import pool from '../db.js';
 
 async function GetExistingSeasonRosterPositions() {
   try {
