+++ services/import/statCategoryService.js
@@ -1,6 +1,6 @@
-import pool from './db.js';
+import pool from '../db.js';
 import yahooApiService from './yahooApiService.js';
-import viewService from './viewService.js';
+import viewService from '../viewService.js';
 import seasonService from './seasonService.js';
 import gamecodeService from './gameCodeService.js';
 import seasonRosterPositionService from './rosterPositionService.js';
